+++ Exporter.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace koolreport\instant;
+
+class Exporter
+{
+    static function export($path)
+    {
+        $report = new ExporterTempReport(array("path"=>$path));
+        return $report->export();
+    }
+}