+++ TempReport.php
@@ -4,5 +4,8 @@ namespace koolreport\instant;
 
 class TempReport extends \koolreport\KoolReport
 {
-
+    function settings()
+    {
+        return $this->params["@reportSettings"];
+    }
 }