+++ Widget.php
@@ -4,20 +4,24 @@ namespace koolreport\instant;
 
 class Widget
 {
-    static function create($class,$params)
+    static function create($widgetClass,$widgetParams,$reportSettings=array())
     {
-        $tempReport = new TempReport(array(
-            "class"=>$class,
-            "params"=>$params,
-        ));
+        $params = array_merge(array(
+            "@widgetClass"=>$widgetClass,
+            "@widgetParams"=>$widgetParams,
+            "@reportSettings"=>$reportSettings
+        ),$_GET,$_POST);
+        $tempReport = new TempReport($params);
         $tempReport->render();
     }
-    static function html($class,$params)
+    static function html($widgetClass,$widgetParams,$reportSettings=array())
     {
-        $tempReport = new TempReport(array(
-            "class"=>$class,
-            "params"=>$params,
-        ));
-        return $tempReport->render(true);        
+        $params = array_merge(array(
+            "@widgetClass"=>$widgetClass,
+            "@widgetParams"=>$widgetParams,
+            "@reportSettings"=>$reportSettings
+        ),$_GET,$_POST);
+        $tempReport = new TempReport($params);
+        return $tempReport->render(true);    
     }
 }