+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# Version 2.0.0
+    1. Bind `$_GET` and `$_POST` to the parameters of KoolReport
+    2. Adding `reportSettings` parameters for temporary report.
+    3. Adding `Exporter` for instant
+
 # Version 1.5.0
     1. Add html() function to widget to return array.